Axis shift obtained to decimal of pixel resolution using NCC slope
Instead of taking the peak of the ifft of the NCC, the slope of the
derivative of the NCC was looked at to determine shift accuracy to a
decimal amount. This plot was frequency shifted to move the noise from
the edges of the image to the sides of the fft plot. The linear fft
correlation plot corresponds to the phase shift. The closest second
order max and min of the NCC from the center was determined to isolate
the linear portion to generate a better polyfit. The slope of the
polyfit line is the estimate of the axis shift between images.